requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
build==1.3.0
certifi==2025.10.5
click==8.3.0
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
Cython==3.1.4
distro==1.9.0
exceptiongroup==1.3.0
Flask @ file:///C:/b/abs_28dbmjidmh/croot/flask_1756453142071/work
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous @ file:///C:/b/abs_c4vwgdr5yn/croot/itsdangerous_1716533399914/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jiter==0.11.0
joblib @ file:///C:/miniconda3/conda-bld/joblib_1757926370033/work
MarkupSafe==3.0.3
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
nltk==3.9.2
numpy==2.2.5
openai==2.1.0
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
pyphen==0.17.2
pyproject_hooks==1.2.0
python-dotenv==1.1.1
regex==2025.9.18
scikit-learn==1.0.2
scipy @ file:///C:/b/abs_71q96u23a4/croot/scipy_1747238057723/work/dist/scipy-1.15.3-cp310-cp310-win_amd64.whl#sha256=06013d45c5cad262bccdca1fbe5c25a720be04ae66dc83d75d1f749d118a7346
sniffio==1.3.1
SQLAlchemy==2.0.43
textstat==0.7.10
threadpoolctl @ file:///C:/b/abs_def0dwqlft/croot/threadpoolctl_1719407816649/work
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
